Update README.md
This project displays the major sorting algorithms such as selection, bubble, insert, quick, and merge sort and compares their runtimes using different input sizes.  The input sizes were given random numbers, ranging from 100-100,000 using a random number generator. To run this C++ program, I used ctime functions to get and display the amount of time it took for each algorithm. From this project, our goal was to understand the Big O concept and came to a conclusion that the merge and quick sort algorithms performed the most efficiently in the shortest amount of time.